// SecondApp/src/Model/EmployeeModel.ts
export class Employee {
	EmpID: number = 0;
	EmpName: string = "";
	EmpEmail: string = "";
	EmpDob: string = "";
	EmpAddress: string = "";
	EmpContact: string = "";
}

const EMPLOYEES: Employee[] = [
  {EmpID:100001, EmpName:"Debraj Dutta", EmpEmail: "a@company.com", EmpDob: "01/01/1983", EmpAddress: "Bangalore", EmpContact: "9836114767"},
  {EmpID:100002, EmpName:"Debraj Dutta", EmpEmail: "a@company.com", EmpDob: "01/01/1983", EmpAddress: "Bangalore", EmpContact: "9836114767"},
  {EmpID:100003, EmpName:"Debraj Dutta", EmpEmail: "a@company.com", EmpDob: "01/01/1983", EmpAddress: "Bangalore", EmpContact: "9836114767"},
  {EmpID:100004, EmpName:"Debraj Dutta", EmpEmail: "a@company.com", EmpDob: "01/01/1983", EmpAddress: "Bangalore", EmpContact: "9836114767"},
  {EmpID:100005, EmpName:"Debraj Dutta", EmpEmail: "a@company.com", EmpDob: "01/01/1983", EmpAddress: "Bangalore", EmpContact: "9836114767"}
];